requirements.txt
viaa-chassis==0.0.4
ldap3==2.6.1 
psycopg2==2.8.4